requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
python-docx
docx2txt